+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changelog
 ------------------
 
 - the `locate_file` method now supports looking for files with different image/annotation prefixes
+- added the `cron` dummy reader which outputs a string according to the provided execution expression
+  (e.g., *every 10 seconds* or *every 5 minutes on workdays*)
 
 
 0.0.1 (2025-10-31)